+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v1.4.1 (2020.2.9)
+
+- Add some popup tips to global vars (#18)
+- New tier guide and UI: reset/rescale GP
+
 ### v1.4.0 (2020.2.6)
 
 - Add switch of showing ilvl. Default closed.